map updtaes try 2
When a user searches a location can the pin on the map be moved to the resulting coordinates? Can the results and glide slide be hidden untill the user clicks "fetch weather data" Also hide them if the coordinates are edited. Also if a person clicks on the map, clear the location search box.