Update index.html
I need to make the following updates to this script. I've tried several times but the formating of the glide slide keeps getting messed up so I have had to revert back several times. When a user searches a location, can the pin on the map be moved to the resulting coordinates?  Also, if a person clicks on the map, can the code automatically clear the location search box?